// src/providers/generated/ups.ts
export interface components {
    schemas: {
// ...
        Activity: {
            /**
             * @description The date of the activity. Format: YYYYMMDD
             * @example 20210210
             */
            date?: string;
            /**
             * @description gmtDate
             * @example 20210210
             */
            gmtDate?: string;
            /**
             * @description gmtOffset
             * @example -05:00
             */
            gmtOffset?: string;
            /**
             * @description gmtTime
             * @example 74700
             */
            gmtTime?: string;
            location?: components["schemas"]["Location"];
            status?: components["schemas"]["Status"];
            /**
             * @description The time of the activity. Format: HHMMSS (24 hr)
             * @example 071356
             */
            time?: string;
        };
// ...
        Package: {
            accessPointInformation?: components["schemas"]["AccessPointInformation"];
            activity?: components["schemas"]["Activity"][];
            /**
             * @description The list of additional attributes that may be associated with the package. Presence of any element indicates the package has that attribute.
             * @example [
             *       "SENSOR_EVENT"
             *     ]
             */
            additionalAttributes?: string[];
            /**
             * @description The list of additional services that may be associated with the package. Presence of any element indicates that the package has that service.
             * @example [
             *       "ADULT_SIGNATURE_REQUIRED",
             *       "SIGNATURE_REQUIRED",
             *       "ADDITIONAL_HANDLING",
             *       "CARBON_NEUTRAL",
             *       "UPS_PREMIER_SILVER",
             *       "UPS_PREMIER_GOLD",
             *       "UPS_PREMIER_PLATINUM"
             *     ]
             */
            additionalServices?: string[];
            alternateTrackingNumber?: components["schemas"]["AlternateTrackingNumber"][];
            currentStatus?: components["schemas"]["Status"];
            deliveryDate?: components["schemas"]["DeliveryDate"][];
            deliveryInformation?: components["schemas"]["DeliveryInformation"];
            deliveryTime?: components["schemas"]["DeliveryTime"];
            dimension?: components["schemas"]["Dimension"];
            /** @description Indicator of whether the package is a smart package */
            isSmartPackage?: boolean;
            /** @description milestones */
            milestones?: components["schemas"]["Milestones"][];
            packageAddress?: components["schemas"]["PackageAddress"][];
            /**
             * Format: int32
             * @description The total number of packages in the shipment. Note that this number may be greater than the number of returned packages in the response. In such cases subsequent calls are needed to get additional packages.
             * @example 2
             */
            packageCount?: number;
            paymentInformation?: components["schemas"]["PaymentInformation"][];
            referenceNumber?: components["schemas"]["ReferenceNumber"][];
            service?: components["schemas"]["Service"];
            statusCode?: string;
            /** @description The activity status description. Note: this field will be translated based on the locale provided in the request. */
            statusDescription?: string;
            /**
             * @description Contains values which signify that certain data should be suppressed or hidden. Valid values: Tracking activity details should be hidden. Note: this is mainly intended for use by UPS.com applications.
             * @example DETAIL
             */
            suppressionIndicators?: string[];
            trackingNumber?: string;
            /** @description This indicator provides UCIX (UPS Customer Information Exchange) status Valid values: 'O' means open. 'C' means closed. */
            ucixStatus?: string;
            weight?: components["schemas"]["Weight"];
        };
// ...
        Response: {
            response?: components["schemas"]["ErrorResponse"];
        };
// ...
        TrackApiResponse: {
            trackResponse?: components["schemas"]["TrackResponse"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}